osint-tracker/requirements.txt
# Core
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# Data processing
pandas>=2.2.0
numpy>=1.26.3
scipy>=1.11.4

# Visualization
folium==0.15.1
rich==13.7.0          # Beautiful terminal output
plotly==5.18.0

# Async HTTP
aiohttp==3.9.1

# Satellite imagery
sentinelhub==3.9.4

# Caching
diskcache==5.6.3

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
responses==0.24.1      # Mock HTTP for tests

# CLI
click==8.1.7

# Date utilities
python-dateutil==2.8.2
